fix: format specifier for scales
